+++ internal/provider/provider.go
@@ -84,14 +84,14 @@ func (p *IncidentProvider) Configure(ctx context.Context, req provider.Configure
 	} else if data.Endpoint.IsNull() || data.Endpoint.IsUnknown() {
 		endpoint = "https://api.incident.io"
 	} else {
-		endpoint = data.Endpoint.String()
+		endpoint = data.Endpoint.ValueString()
 	}
 
 	var apiKey string
 	if data.APIKey.IsNull() || data.APIKey.IsUnknown() {
 		apiKey = os.Getenv("INCIDENT_API_KEY")
 	} else {
-		apiKey = data.APIKey.String()
+		apiKey = data.APIKey.ValueString()
 	}
 
 	bearerTokenProvider, bearerTokenProviderErr := securityprovider.NewSecurityProviderBearerToken(apiKey)