+++ internal/provider/incident_catalog_type_resource.go
@@ -1,0 +1,191 @@
+package provider
+
+import (
+	"context"
+	"fmt"
+
+	"github.com/hashicorp/terraform-plugin-framework/path"
+	"github.com/hashicorp/terraform-plugin-framework/resource"
+	"github.com/hashicorp/terraform-plugin-framework/resource/schema"
+	"github.com/hashicorp/terraform-plugin-framework/resource/schema/planmodifier"
+	"github.com/hashicorp/terraform-plugin-framework/resource/schema/stringplanmodifier"
+	"github.com/hashicorp/terraform-plugin-framework/types"
+	"github.com/hashicorp/terraform-plugin-log/tflog"
+	"github.com/incident-io/terraform-provider-incident/internal/apischema"
+	"github.com/incident-io/terraform-provider-incident/internal/client"
+)
+
+var (
+	_ resource.Resource                = &IncidentCatalogTypeResource{}
+	_ resource.ResourceWithImportState = &IncidentCatalogTypeResource{}
+)
+
+type IncidentCatalogTypeResource struct {
+	client *client.ClientWithResponses
+}
+
+type IncidentCatalogTypeResourceModel struct {
+	ID           types.String `tfsdk:"id"`
+	Name         types.String `tfsdk:"name"`
+	Description  types.String `tfsdk:"description"`
+	SemanticType types.String `tfsdk:"semantic_type"`
+}
+
+func NewIncidentCatalogTypeResource() resource.Resource {
+	return &IncidentCatalogTypeResource{}
+}
+
+func (r *IncidentCatalogTypeResource) Metadata(ctx context.Context, req resource.MetadataRequest, resp *resource.MetadataResponse) {
+	resp.TypeName = req.ProviderTypeName + "_catalog_type"
+}
+
+func (r *IncidentCatalogTypeResource) Schema(ctx context.Context, req resource.SchemaRequest, resp *resource.SchemaResponse) {
+	resp.Schema = schema.Schema{
+		MarkdownDescription: apischema.TagDocstring("Catalog V2"),
+		Attributes: map[string]schema.Attribute{
+			"id": schema.StringAttribute{
+				Computed:            true,
+				MarkdownDescription: apischema.Docstring("CatalogTypeV2ResponseBody", "id"),
+				PlanModifiers: []planmodifier.String{
+					stringplanmodifier.UseStateForUnknown(),
+				},
+			},
+			"name": schema.StringAttribute{
+				MarkdownDescription: apischema.Docstring("CatalogV2CreateTypeRequestBody", "name"),
+				Required:            true,
+			},
+			"description": schema.StringAttribute{
+				MarkdownDescription: apischema.Docstring("CatalogV2CreateTypeRequestBody", "description"),
+				Required:            true,
+			},
+			"semantic_type": schema.StringAttribute{
+				MarkdownDescription: apischema.Docstring("CatalogV2CreateTypeRequestBody", "semantic_type"),
+				Computed:            true,
+				Optional:            true,
+			},
+		},
+	}
+}
+
+func (r *IncidentCatalogTypeResource) Configure(ctx context.Context, req resource.ConfigureRequest, resp *resource.ConfigureResponse) {
+	if req.ProviderData == nil {
+		return
+	}
+
+	client, ok := req.ProviderData.(*client.ClientWithResponses)
+	if !ok {
+		resp.Diagnostics.AddError(
+			"Unexpected Resource Configure Type",
+			fmt.Sprintf("Expected *client.Client, got: %T. Please report this issue to the provider developers.", req.ProviderData),
+		)
+
+		return
+	}
+
+	r.client = client
+}
+
+func (r *IncidentCatalogTypeResource) Create(ctx context.Context, req resource.CreateRequest, resp *resource.CreateResponse) {
+	var data *IncidentCatalogTypeResourceModel
+	resp.Diagnostics.Append(req.Plan.Get(ctx, &data)...)
+	if resp.Diagnostics.HasError() {
+		return
+	}
+
+	var semanticType string
+	if data.SemanticType.IsUnknown() {
+		semanticType = "custom"
+	} else {
+		semanticType = data.SemanticType.ValueString()
+	}
+	result, err := r.client.CatalogV2CreateTypeWithResponse(ctx, client.CreateTypeRequestBody{
+		Name:         data.Name.ValueString(),
+		Description:  data.Description.ValueString(),
+		SemanticType: semanticType,
+	})
+	if err == nil && result.StatusCode() >= 400 {
+		err = fmt.Errorf(string(result.Body))
+	}
+	if err != nil {
+		resp.Diagnostics.AddError("Client Error", fmt.Sprintf("Unable to create catalog type, got error: %s", err))
+		return
+	}
+
+	tflog.Trace(ctx, fmt.Sprintf("created a catalog type resource with id=%s", result.JSON201.CatalogType.Id))
+	data = r.buildModel(result.JSON201.CatalogType)
+	resp.Diagnostics.Append(resp.State.Set(ctx, &data)...)
+}
+
+func (r *IncidentCatalogTypeResource) Read(ctx context.Context, req resource.ReadRequest, resp *resource.ReadResponse) {
+	var data *IncidentCatalogTypeResourceModel
+	resp.Diagnostics.Append(req.State.Get(ctx, &data)...)
+	if resp.Diagnostics.HasError() {
+		return
+	}
+
+	result, err := r.client.CatalogV2ShowTypeWithResponse(ctx, data.ID.ValueString())
+	if err != nil {
+		resp.Diagnostics.AddError("Client Error", fmt.Sprintf("Unable to read catalog type, got error: %s", err))
+		return
+	}
+
+	data = r.buildModel(result.JSON200.CatalogType)
+	resp.Diagnostics.Append(resp.State.Set(ctx, &data)...)
+}
+
+func (r *IncidentCatalogTypeResource) Update(ctx context.Context, req resource.UpdateRequest, resp *resource.UpdateResponse) {
+	var data *IncidentCatalogTypeResourceModel
+	resp.Diagnostics.Append(req.Plan.Get(ctx, &data)...)
+	if resp.Diagnostics.HasError() {
+		return
+	}
+
+	var semanticType string
+	if data.SemanticType.IsUnknown() {
+		semanticType = "custom"
+	} else {
+		semanticType = data.SemanticType.ValueString()
+	}
+	result, err := r.client.CatalogV2UpdateTypeWithResponse(ctx, data.ID.ValueString(), client.CatalogV2UpdateTypeJSONRequestBody{
+		Name:         data.Name.ValueString(),
+		Description:  data.Description.ValueString(),
+		SemanticType: semanticType,
+	})
+	if err == nil && result.StatusCode() >= 400 {
+		err = fmt.Errorf(string(result.Body))
+	}
+	if err != nil {
+		resp.Diagnostics.AddError("Client Error", fmt.Sprintf("Unable to update catalog type, got error: %s", err))
+		return
+	}
+
+	data = r.buildModel(result.JSON200.CatalogType)
+	resp.Diagnostics.Append(resp.State.Set(ctx, &data)...)
+}
+
+func (r *IncidentCatalogTypeResource) Delete(ctx context.Context, req resource.DeleteRequest, resp *resource.DeleteResponse) {
+	var data *IncidentCatalogTypeResourceModel
+	resp.Diagnostics.Append(req.State.Get(ctx, &data)...)
+	if resp.Diagnostics.HasError() {
+		return
+	}
+
+	_, err := r.client.CatalogV2DestroyTypeWithResponse(ctx, data.ID.ValueString())
+	if err != nil {
+		resp.Diagnostics.AddError("Client Error", fmt.Sprintf("Unable to delete catalog type, got error: %s", err))
+		return
+	}
+}
+
+func (r *IncidentCatalogTypeResource) ImportState(ctx context.Context, req resource.ImportStateRequest, resp *resource.ImportStateResponse) {
+	resource.ImportStatePassthroughID(ctx, path.Root("id"), req, resp)
+}
+
+func (r *IncidentCatalogTypeResource) buildModel(catalogType client.CatalogTypeV2) *IncidentCatalogTypeResourceModel {
+	return &IncidentCatalogTypeResourceModel{
+		ID:           types.StringValue(catalogType.Id),
+		Name:         types.StringValue(catalogType.Name),
+		Description:  types.StringValue(catalogType.Description),
+		SemanticType: types.StringValue(catalogType.SemanticType),
+	}
+}